requirements.txt
alembic==1.8.1
appnope==0.1.2
astroid==2.9.3
attrs==21.2.0
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
debugpy==1.4.3
decorator==5.1.0
entrypoints==0.3
Flask==2.1.3
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.1.2
jupyter-client==7.0.5
jupyter-core==4.8.1
lazy-object-proxy==1.7.1
Mako==1.2.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.6.1
nest-asyncio==1.5.1
packaging==20.9
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.1
pluggy==0.13.1
prompt-toolkit==3.0.20
psycopg2-binary==2.9.3
ptyprocess==0.7.0
py==1.10.0
Pygments==2.10.0
pylint==2.12.2
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
python-dotenv==0.20.0
pyzmq==22.3.0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.39
toml==0.10.2
tomli==2.0.1
tornado==6.1
traitlets==5.1.0
typing_extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.2.0
wonderwords==2.2.0
wrapt==1.13.3
zipp==3.8.1
